updated fixes for vercel deployment
